SoundMgr 및 Sound파일추가

+++ Project_C/Assets/Script/Common/Character/Character_Critical.cs
@@ -13,7 +13,7 @@ public abstract partial class Character
 
         if(isCritical)
         {
-            _damagedata.Damage *= CiritcalMultiplier;
+            _damagedata.Damage *= (int)CiritcalMultiplier;
             _damagedata.DamageFont_Type = DAMAGEFONT_TYPE.CRITICAL;
         }
     }
