+++ Project_C/Assets/Script/Common/StatData/StatData.cs
@@ -13,7 +13,7 @@ public class StatData : MonoBehaviour
 
     public void TakeDamage(DamageData _damagedata)
     {
-        _damagedata.Damage = _damagedata.Damage - dicstat[STAT_TYPE.DEF];
+        _damagedata.Damage = _damagedata.Damage - (int)dicstat[STAT_TYPE.DEF];
         Shared.GameMgr.DAMAGE_FONT.CreateDamageFont(_damagedata, uihead.position);
         dicstat[STAT_TYPE.HP] = dicstat[STAT_TYPE.HP] - _damagedata.Damage;
     }
